requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.2
    # via langchain-community
aiosignal==1.4.0
    # via aiohttp
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   httpx
    #   starlette
asttokens==3.0.1
    # via stack-data
attrs==25.4.0
    # via aiohttp
certifi==2025.11.12
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via uvicorn
dataclasses-json==0.6.7
    # via langchain-community
decorator==5.2.1
    # via ipython
executing==2.2.1
    # via stack-data
fastapi==0.121.2
    # via -r requirements.in
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
greenlet==3.2.4
    # via sqlalchemy
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   langgraph-sdk
    #   langsmith
httpx-sse==0.4.3
    # via langchain-community
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
ipython==9.7.0
    # via -r requirements.in
ipython-pygments-lexers==1.1.1
    # via ipython
jedi==0.19.2
    # via ipython
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==1.0.7
    # via -r requirements.in
langchain-classic==1.0.0
    # via langchain-community
langchain-community==0.4.1
    # via -r requirements.in
langchain-core==1.0.5
    # via
    #   -r requirements.in
    #   langchain
    #   langchain-classic
    #   langchain-community
    #   langchain-text-splitters
    #   langgraph
    #   langgraph-checkpoint
    #   langgraph-prebuilt
langchain-text-splitters==1.0.0
    # via langchain-classic
langgraph==1.0.3
    # via
    #   -r requirements.in
    #   langchain
langgraph-checkpoint==3.0.1
    # via
    #   langgraph
    #   langgraph-prebuilt
langgraph-prebuilt==1.0.4
    # via langgraph
langgraph-sdk==0.2.9
    # via langgraph
langsmith==0.4.43
    # via
    #   langchain-classic
    #   langchain-community
    #   langchain-core
lxml==6.0.2
    # via pandas-datareader
marshmallow==3.26.1
    # via dataclasses-json
matplotlib-inline==0.2.1
    # via ipython
multidict==6.7.0
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.1.0
    # via typing-inspect
numpy==2.3.4
    # via
    #   langchain-community
    #   pandas
orjson==3.11.4
    # via
    #   langgraph-sdk
    #   langsmith
ormsgpack==1.12.0
    # via langgraph-checkpoint
packaging==25.0
    # via
    #   langchain-core
    #   langsmith
    #   marshmallow
pandas==2.3.3
    # via pandas-datareader
pandas-datareader==0.10.0
    # via -r requirements.in
parso==0.8.5
    # via jedi
pexpect==4.9.0
    # via ipython
prompt-toolkit==3.0.52
    # via ipython
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.3
    # via stack-data
pydantic==2.12.4
    # via
    #   -r requirements.in
    #   fastapi
    #   langchain
    #   langchain-classic
    #   langchain-core
    #   langgraph
    #   langsmith
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.12.0
    # via langchain-community
pygments==2.19.2
    # via
    #   ipython
    #   ipython-pygments-lexers
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.2.1
    # via pydantic-settings
pytz==2025.2
    # via pandas
pyyaml==6.0.3
    # via
    #   langchain-classic
    #   langchain-community
    #   langchain-core
requests==2.32.5
    # via
    #   langchain-classic
    #   langchain-community
    #   langsmith
    #   pandas-datareader
    #   requests-toolbelt
requests-toolbelt==1.0.0
    # via langsmith
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.44
    # via
    #   langchain-classic
    #   langchain-community
stack-data==0.6.3
    # via ipython
starlette==0.49.3
    # via fastapi
tenacity==9.1.2
    # via
    #   langchain-community
    #   langchain-core
traitlets==5.14.3
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.15.0
    # via
    #   aiosignal
    #   anyio
    #   fastapi
    #   langchain-core
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   starlette
    #   typing-inspect
    #   typing-inspection
typing-inspect==0.9.0
    # via dataclasses-json
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via requests
uvicorn==0.38.0
    # via -r requirements.in
wcwidth==0.2.14
    # via prompt-toolkit
xxhash==3.6.0
    # via langgraph
yarl==1.22.0
    # via aiohttp
zstandard==0.25.0
    # via langsmith
